alx_travel_app/requirements.txt
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
Faker==19.13.0
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4